# swagger.json
{
  "swagger": "2.0",
    "info": {
      "description":"THIS IS MY BRAND API ",
      "version": "1.0.0",
      "title": "BRAND-API",
      "contact": {
        "email": "igihozoinnocente@gmail.com"
      }
    },
  "servers": [

    {

      "url": "http://localhost:3000",

      "name": "DEV"

    },

    {

      "url": "https://staging-ft-auth-validat-qg7m2y.herokuapp.com/",

      "name": "PROD"

    }

  ],
  "paths": {
    "/api/v1/users/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create user",
        "description": "This can only be done by a new user.",
        "operationId": "createUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/api/v1/articles/{id}": {

      "get": {

        "tags": ["Articles"],

        "summary": "Get one article",

        "description": "Get one article",

        "operationId": "getoneArticle",

        "parameters": [

          {

            "name": "id",

            "in": "path",

            "description": "id",

            "required": true

          }

        ],

        "responses": {

          "200": {

            "description": "Success"

          }

        }

      },

      "post": {

        "tags": ["Articles"],

        "summary": "Create article",

        "description": "Create article",

        "operationId": "postArticles",

        "parameters": [

          {

            "name": "Authorization",

            "in": "header",

            "description": "Authorization",

            "required": true

          },

          {

            "name": "title",

            "in": "formData",

            "description": "title",

            "required": true

          },

          {

            "name": "content",

            "in": "formData",

            "description": "content",

            "required": true

          },

          {

            "name": "image",

            "in": "formData",

            "type": "file",

            "description": "image",

            "required": true

          }

        ],

        "responses": {

          "200": {

            "description": "Success"

          }

        }

      },

      "patch": {

        "tags": ["Articles"],

        "summary": "Update article",

        "description": "Update article",

        "operationId": "updateArticle",

        "parameters": [

          {

            "name": "Authorization",

            "in": "header",

            "description": "Authorization",

            "required": false

          },

          {

            "name": "id",

            "in": "path",

            "description": "id",

            "required": true

          },

          {

            "name": "title",

            "in": "formData",

            "description": "title",

            "required": false

          },

          {

            "name": "content",

            "in": "formData",

            "description": "content",

            "required": false

          },

          {

            "name": "image",

            "in": "formData",

            "type": "file",

            "description": "image",

            "required": false

          }

        ],

        "responses": {

          "200": {

            "description": "Success"

          }

        }

      },

      "delete": {

        "tags": ["Articles"],

        "summary": "Delete one article",

        "description": "Delete one article",

        "operationId": "deleteoneArticle",

        "parameters": [

          {

            "name": "Authorization",

            "in": "header",

            "description": "Authorization",

            "required": true

          },

          {

            "name": "id",

            "in": "path",

            "description": "id",

            "required": true

          }

        ],

        "responses": {

          "200": {

            "description": "Success"

          }

        }

      }

    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  }
}